backend/requirements.txt
 # Core dependencies
flask>=2.0.0
flask-cors>=3.0.0
flask-sqlalchemy>=3.0.0
werkzeug>=2.0.0
pyjwt>=2.0.0
python-dotenv>=0.19.0

# Database
sqlalchemy>=2.0.0
alembic>=1.10.0

# Vector storage
weaviate-client>=4.0.0

# LLM integration
langchain>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0

# Utilities
requests>=2.25.0
python-jose>=3.3.0
uuid>=1.30